Fix weird bug with data packing

+++ src/data.h
@@ -6,7 +6,7 @@
 
 void data_pack_int32(unsigned char*, int32_t);
 void data_pack_int16(unsigned char*, int16_t);
-void data_pack_float32(unsigned char*, float);
+void data_pack_float32(unsigned char*, double);
 
 int32_t data_unpack_int32(unsigned char*);
 int16_t data_unpack_int16(unsigned char*);
